Split HIR model definitions (#1298)
Move the public HIR structs, enums, type relation helpers, and display methods from hir.rs into hir/model.rs while preserving the crate::hir re-export surface. Update the monolith reporter and decomposition ratchets for the lower HIR facade line count.

Refs #1254.
